lesto

+++ funciones.py
@@ -3,36 +3,36 @@ import numpy as np
 
 def unidad(x):
     unidad.__name__ = '1'
-    return 1.0
+    return np.float64(1)
 
 
 def inden(x):
     inden.__name__ = 'x'
-    return float(x)
+    return np.float64(x)
 
 
 def x_pow(p=1):
     x_pow.__name__ = 'x^'+str(p)
-    return lambda x: x**p
+    return lambda x: np.float64(x**p)
 
 
 def sen(x):
     sen.__name__ = 'sen(x)'
-    return float(np.sin(x))
+    return np.sin(x)
 
 
 def cos(x):
     cos.__name__ = 'cos(x)'
-    return float(np.cos(x))
+    return np.cos(x)
 
 
 def exp(n=False):
     if n:
-        return lambda x: float(np.exp(-x))
+        return lambda x: np.exp(-x)
     else:
-        return lambda x: float(np.exp(x))
+        return lambda x: np.exp(x)
 
 
 def ln(x):
     ln.__name__ = 'ln(x)'
-    return float(np.log(x))
+    return np.log(x)
